+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changed
   * AZ-398: Force lowercases on default generated name
 
 Added
-  * AZ-365 Add diagnostics settings
+  * AZ-365: Add diagnostics settings
   
 # v3.1.0 - 2020-08-25